Add migration and usage guides
Since this new version of the extension contains changes which affect
existing users of the extension, this commit introduces a migration
guide which opens only if you upgrade from the original version of the
extension. This migration page describes clearly that the extension has
changed in a way that will affect existing users and provides a button
to request permissions for the Google Meet and Hangouts sites so that
they can be enabled by default.

For those who are installing the extension for the first time, a usage
guide will appear.