Expand support for more websites
This commit adds support for more websites (such as MS Teams) by
patching the deprecated "getUserMedia" function on the `navigator`
object (and its vendor-prefixed equivalents). Additionally, it also now
patches the MediaStreamTrack's applyConstraints function. The patched
function now sets the `autoGainControl` property of the constraints to
false, even if `autoGainControl` was not explicitly requested (set to
`true`). This was also required to support Microsoft Teams (and I
suspect several other sites), since it does not explicitly set the
`autoGainControl` property on the audio constraints to `true`.